Renamed tests (local analyze not working)
