+++ packages/flutter/test/painting/network_image_web_test.dart
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import '_network_image_web_io_test.dart'
-    if (dart.library.html) '_network_image_web_test.dart';
+import '_network_image_web_test_io.dart'
+    if (dart.library.html) '_network_image_test_web.dart';
 
 void main() {
   runTests();
